separate out Socean class, add helpers.ts, types.ts
