+++ src/helpers.ts
@@ -1,0 +1,20 @@
+import { PublicKey } from "@solana/web3.js";
+
+export function reverse(object: any) {
+  for (const val in object) {
+    if (object[val] instanceof PublicKey) {
+      object[val] = new PublicKey(object[val].toBytes().reverse());
+      //console.log(val, object[val].toString());
+    } else if (object[val] instanceof Object) {
+      reverse(object[val]);
+    } else if (object[val] instanceof Array) {
+      for (const elem of object[val]) {
+        reverse(elem);
+      }
+    }
+    /*else {
+      console.log(val, object[val]);
+    }*/
+  }
+}
+
